Deleted ExternalChartAveragesCreator - used code from it in ChartAveragesCreator
