+++ app/src/main/java/pl/llp/aircasting/ui/view/screens/dashboard/charts/ExternalSessionChartDataCalculator.kt
@@ -25,7 +25,7 @@ class ExternalSessionChartDataCalculator(session: Session) : SessionChartDataCal
                 this::setStartEndTimeToDisplay
             )
 
-            else -> ExternalChartAveragesCreator().getFixedEntries(
+            else -> ChartAveragesCreator().getFixedEntries(
                 stream,
                 this::setStartEndTimeToDisplay
             )
