+++ app/src/main/java/pl/llp/aircasting/ui/view/screens/dashboard/charts/OpenAQpmChartAveragesCreator.kt
@@ -2,7 +2,7 @@ package pl.llp.aircasting.ui.view.screens.dashboard.charts
 
 import java.util.*
 
-class OpenAQpmChartAveragesCreator : ExternalChartAveragesCreator() {
+class OpenAQpmChartAveragesCreator : ChartAveragesCreator() {
     override fun modifyHours(date: Date, hours: Int): Date {
         return super.modifyHours(date, -2)
     }
